requirements.txt
Django==1.11.1
django-bulk-update==2.2.0
django-celery==3.1.17
django-cors-headers==1.1.0
django-extensions==1.6.1
django-filter==0.12.0
django-grappelli==2.10.1
django-import-export==0.4.2
django-model-utils==3.1.1
django-nose==1.4.5
django-reversion==2.0.13
django-test-without-migrations==0.6
djangorestframework==3.1.3
MySQL-python==1.2.5
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.8
pkg-resources==0.0.0
